refactor: paths module
